server: fix a segfault when a broken client logs in with color 0
When the construction of the player fails, we do not want a half-made instance
to remain in the player list. It can happen if a client asks to join with an
erroneous color 0. The error handling frees memory, but last_player->next will
already point there. The result is a server segfault. We could fix it by
setting the pointer back to NULL, but a cleaner fix is to only add fully
constructed objects on the list.

Remove the odd macro and add the new player to the list only when it's fully
constructed.